Added basic obstacle generation, removed obstacles file
